requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
pymongo==4.6.1
python-dotenv==1.0.0
eventlet==0.33.3
